Refactor: move API access logging module
