+++ app.js
@@ -7,11 +7,11 @@ const cookieParser = require("cookie-parser");
 const expressSession = require("express-session");
 const MongoStore = require("connect-mongo");
 const cors = require("cors");
-const responseTimeMiddleware = require("response-time");
 
 // 내부 모듈
 const security = require("./security");
 const logger = require("./src/modules/logger");
+const logAPIAccess = require("./src/modules/logAPIAccess");
 const startSocketServer = require("./src/modules/socket");
 
 // 익스프레스 서버 생성
@@ -33,7 +33,7 @@ app.use(session);
 app.use(cookieParser());
 
 // API 접근 기록 및 응답 시간을 http response의 헤더에 기록합니다.
-app.use(responseTimeMiddleware(require("./src/modules/apiAccessLog")));
+app.use(require("response-time")(logAPIAccess));
 
 // 라우터 및 리액트
 app.use("/auth", require("./src/route/auth"));
